feat(app): Great progress on fixing the app
env example file removed because it's unnecessary, added app id to bot initialization expression, created setup functions to aid in loading functions across modules, added the capability to dynamically load command modules based on a list, this seems to be working but will probably need to be cleaned up later, added async and better checking to ensure the bot loads as intended with exception handling, added a markdown file containing the current command progress, this will be added to when additional progress is made. all commands had modifications to allow them to be added to the bot without being in the same file, this caused a lot of downstream effects and is still being worked on, see the progress markdown file for additional details. the other remaining files had changes to facilitate patching issues arising from the refactor.